pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [{ name = "Alvaro Bartolome", email = "alvarobartt@gmail.com" }]
classifiers = [
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "docker",
    "google-api-core",
    "google-cloud-aiplatform",
    "google-cloud-artifact-registry",
    "google-cloud-storage",
    "huggingface_hub[hf_transfer]",
    "rich",
]
description = "🤗 Hugging Face Inference Toolkit for Google Cloud Vertex AI (similar to SageMaker's Inference Toolkit, but unofficial)"
dynamic = ["version"]
keywords = []
license = "Apache-2.0"
name = "vertex_ai_huggingface_inference_toolkit"
readme = "README.md"
requires-python = ">=3.8"

[project.urls]
Documentation = "https://alvarobartt.github.io/vertex-ai-huggingface-inference-toolkit"
Issues = "https://github.com/alvarobartt/vertex-ai-huggingface-inference-toolkit/issues"
Source = "https://github.com/alvarobartt/vertex-ai-huggingface-inference-toolkit"

[tool.hatch.version]
path = "src/vertex_ai_huggingface_inference_toolkit/__init__.py"

[project.optional-dependencies]
transformers = ["accelerate", "transformers"]
diffusers = ["accelerate", "diffusers"]
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocstrings[python]",
]
quality = ["mypy~=1.4.0", "ruff>=0.3.0", "pre-commit>=3.5.0"]
tests = ["pytest~=7.1.2", "pytest-cov~=4.1"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
warn_return_any = true

[tool.ruff]
line-length = 88
exclude = ["docs"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B"]
ignore = ["E501", "B008", "C901"]

[tool.ruff.lint.isort]
known-first-party = ["vertex_ai_huggingface_inference_toolkit"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "/.vscode",
    "/docs",
    "/.pre-commit-config.yaml",
    "/.gitignore",
    "/tests",
]
